# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: The Obsidian Curator API
  version: 0.1.0
  description: >-
    Phase 0 OpenAPI scaffold for the production e-commerce transformation.
servers:
  - url: http://localhost:4000
    description: Local development
paths:
  /health:
    get:
      summary: API health check
      operationId: getApiHealth
      responses:
        '200':
          description: Healthy
  /api/v1/health:
    get:
      summary: Versioned health check
      operationId: getVersionedApiHealth
      responses:
        '200':
          description: Healthy
  /api/v1/auth/register:
    post:
      summary: Register a customer account
      operationId: register
      responses:
        '201':
          description: Account registered
  /api/v1/auth/login:
    post:
      summary: Login with email and password
      operationId: login
      responses:
        '200':
          description: Login success
  /api/v1/auth/logout:
    post:
      summary: Logout current session
      operationId: logout
      responses:
        '200':
          description: Logout success
  /api/v1/auth/refresh:
    post:
      summary: Refresh auth tokens
      operationId: refreshAuth
      responses:
        '200':
          description: Token refresh success
  /api/v1/auth/forgot-password:
    post:
      summary: Request password reset
      operationId: forgotPassword
      responses:
        '200':
          description: Reset accepted
  /api/v1/auth/reset-password:
    post:
      summary: Reset password with token
      operationId: resetPassword
      responses:
        '200':
          description: Password reset success
  /api/v1/categories:
    get:
      summary: List active categories
      operationId: listCategories
      responses:
        '200':
          description: Categories returned
  /api/v1/products:
    get:
      summary: List published products with filter query support
      operationId: listProducts
      responses:
        '200':
          description: Products returned
  /api/v1/products/{id}:
    get:
      summary: Get product by id or slug
      operationId: getProduct
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product returned
        '404':
          description: Product not found
  /api/v1/products/{id}/related:
    get:
      summary: Get related products by category
      operationId: listRelatedProducts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Related products returned
        '404':
          description: Product not found
  /api/v1/cart:
    get:
      summary: Get authenticated user cart
      operationId: getCart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart returned
        '401':
          description: Unauthorized
  /api/v1/cart/items:
    post:
      summary: Add item to authenticated user cart
      operationId: addCartItem
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart item added
        '400':
          description: Invalid request
        '404':
          description: Product not found
        '409':
          description: Inventory or cart conflict
        '401':
          description: Unauthorized
  /api/v1/cart/items/{itemId}:
    patch:
      summary: Update item quantity in cart
      operationId: updateCartItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cart item updated
        '404':
          description: Cart item not found
        '409':
          description: Inventory or variant conflict
        '401':
          description: Unauthorized
    delete:
      summary: Remove item from cart
      operationId: removeCartItem
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Cart item removed
        '404':
          description: Cart item not found
        '401':
          description: Unauthorized
  /api/v1/checkout/quote:
    post:
      summary: Create checkout quote from cart
      operationId: createCheckoutQuote
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Quote returned
        '400':
          description: Cart is empty
        '401':
          description: Unauthorized
  /api/v1/checkout/intent:
    post:
      summary: Create checkout payment intent
      operationId: createCheckoutIntent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Existing intent returned via idempotency key or concurrent unique conflict recovery
        '201':
          description: Payment intent created and awaiting provider confirmation
        '400':
          description: Cart is empty
        '401':
          description: Unauthorized
  /api/v1/checkout/confirm:
    post:
      summary: Finalize checkout after payment verification
      operationId: confirmCheckout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paid order created
        '400':
          description: Cart is empty or invalid
        '404':
          description: Payment intent not found
        '409':
          description: Cart, inventory, or payment intent conflict
        '401':
          description: Unauthorized
  /api/v1/webhooks/payments/stripe:
    post:
      summary: Handle Stripe payment status callback
      operationId: handleStripePaymentWebhook
      parameters:
        - in: header
          name: x-stripe-signature
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Webhook accepted and payment intent status transitioned
        '401':
          description: Invalid webhook signature
        '404':
          description: Payment intent not found or transition not allowed
  /api/v1/system/meta:
    get:
      summary: API metadata endpoint
      operationId: getSystemMeta
      responses:
        '200':
          description: Metadata returned
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas: {}